Move uniswap v2 staking to its own directory
This allows other staking protocols to live in adjacent directories
