+++ src/components/themed/TransactionListTop.tsx
@@ -15,7 +15,7 @@ import { useWalletName } from '../../hooks/useWalletName'
 import { useWatch } from '../../hooks/useWatch'
 import { formatNumber } from '../../locales/intl'
 import s from '../../locales/strings'
-import { StakePolicy } from '../../plugins/stake-plugins'
+import { StakePolicy } from '../../plugins/stake-plugins/uniswapV2/uniV2Plugin'
 import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
 import { getExchangeRate } from '../../selectors/WalletSelectors'
 import { useDispatch, useSelector } from '../../types/reactRedux'
