+++ src/plugins/stake-plugins/uniswapV2/abi/ANYSWAP_V5_ERC20_ABI.json
