+++ src/plugins/stake-plugins/uniswapV2/abi/MASONRY_ABI.json
