+++ src/plugins/stake-plugins/uniswapV2/abi/TSHARE_WFTM_LP_ABI.json
