+++ src/plugins/stake-plugins/uniswapV2/abi/UNISWAP_V2_ROUTER_02.json
