+++ src/plugins/stake-plugins/uniswapV2/abi/WRAPPED_FTM_ABI.json
