+++ src/plugins/stake-plugins/uniswapV2/policies/cemeteryPolicy.ts
@@ -4,17 +4,17 @@ import { add, div, gt, gte, lte, mul, sub } from 'biggystring'
 import { ethers } from 'ethers'
 import { sprintf } from 'sprintf-js'
 
-import s from '../../../locales/strings'
+import s from '../../../../locales/strings'
+import { cacheTxMetadata } from '../../metadataCache'
+import { AssetId, ChangeQuote, ChangeQuoteRequest, PositionAllocation, QuoteAllocation, StakePosition, StakePositionRequest } from '../../types'
+import { makeBigAccumulator } from '../../util/accumulator'
+import { round } from '../../util/biggystringplus'
+import { makeBuilder } from '../../util/builder'
+import { fromHex } from '../../util/hex'
 import { getContractInfo, makeContract, makeSigner, multipass } from '../contracts'
-import { cacheTxMetadata } from '../metadataCache'
 import { pluginInfo } from '../pluginInfo'
 import { StakePolicyInfo } from '../stakePolicy'
-import { AssetId, ChangeQuote, ChangeQuoteRequest, PositionAllocation, QuoteAllocation, StakePosition, StakePositionRequest } from '../types'
-import { makeBigAccumulator } from '../util/accumulator'
-import { round } from '../util/biggystringplus'
-import { makeBuilder } from '../util/builder'
-import { fromHex } from '../util/hex'
-import { StakePluginPolicy } from './types'
+import { StakePluginPolicy } from '../types'
 
 export interface CemeteryPolicyOptions {
   disableStake?: boolean
